`slice` rationale is to focus on the success category
The old description for why we're using slice made me confuse myself -- it made it sound like one row would be "men - women" and the other would be "women - men."

It's actually that we're finding $\hat{p}_\text{men, promoted} - \hat{p}_\text{women, promoted}$ in the first row, and $\hat{p}_\text{men, not promoted} - \hat{p}_\text{women, not promoted}$ in the second row. That is, the order of the subtraction doesn't change, we're just looking at the second row of the contingency table.

Then it made more sense in my brain why we should slice out the first row and discard the second (we don't care about not-promotion rates).